cr50: change the power button handling
Enable permanently the power button interrupt and discard the unneeded
presses, rather than enabling the interrupt on demand.
This is a preparatory work for the U2F feature which is using the power
button as the user physical presence sensing. It will need to be able to
detect all the button touches.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=cr50
BUG=b:35545754
RUN=with follow-up CLs, run U2FTest on Eve.

Change-Id: Ifd90970932684d31ad7687db260bafa65a56e854