Update patch set 6
Patch Set 6:

This patch also fixes a problem we see on Eve when attempting to boot the EC under a deeply discharged battery condition. In this case once the input current limit is set to 512 mA, PPVAR_VSYS drops, sometimes low enough to pull down the PP3300_DSW rail which causes the EC to reset. Depending on how deeply discharged the battery is, it may reset once and not come back up at all until a the charger is removed and reinserted, or it may cycle through a number of resets until the battery is charged sufficiently. I think this fix should go in.

Patch-set: 6
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>