cr50: usb_upgrade: pass proper number of bytes to the vendor commands
The code invoking vendor commands callbacks rightly passes the pointer
to the command payload as the address right after the subcommand
field, but does not deduct the size of the subcommand field from the
size of the payload passed to the handler.

This patch fixes the issue, the command handlers do not see two extra
bytes at the tail of the command any more.

BRANCH=cr50
BUG=b:62294740, b:35545754
TEST=verified that vendor commands sent over USB and TPM still work
      properly (in particular the TURN_UPDATE_ON command).

Change-Id: I370e5b198022e153ff47f212e0e59a499bfd77d5
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: d0ee126b4cdc368c36ae6660d66fed1524476e59
Original-Change-Id: I11a45f65163044f808a82b214f9c5faf775f9020
Original-Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/522943
Original-Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>