Update patch set 3
Patch Set 3:

I am admittedly very rusty on this subject, so this is mostly for my education:

I was under impression that there should be only a single context using dcrypto hardware, and the protection was at the operation initialization time, i.e. the first caller wins the hardware and holds it until the operation is completed. The rest of the callers use SW contexts and can coexist.

Is this not the case any more?

Patch-set: 3
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>