Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > I am admittedly very rusty on this subject, so this is mostly for my education:
> > > 
> > > I was under impression that there should be only a single context using dcrypto hardware, and the protection was at the operation initialization time, i.e. the first caller wins the hardware and holds it until the operation is completed. The rest of the callers use SW contexts and can coexist.
> > > 
> > > Is this not the case any more?
> > 
> > This is exactly the information that I don't have. If somebody could definitively answer this question, I would be grateful.
> 
> +1

This change itself is fine but raises the question of what the the expectations of your higher level design are. Hw resources are state and sharing state between threads is fraught w/ peril. A (mental) model where a single task ("driver") is in charge of handling the engine and provides an RPC for others to peruse is preferable. That'd give us a place to stand to try leverage the memory protection unit to enforce this separation. Currently, we're somewhere in the undefined middle.

Another model is to view the engine as a closure executor and always go through a full grab/load code/load data/execute/wipe cycle/release.

There are no feasible run-time fallbacks for asymmetric crypto in case the engine is busy (unlike for the simpler hashing operations).

But yes, DONE interrupts are expected to have matching start counterparts. The interface between the m3 and the engine is a pretty simple state machine w/ handshake but has no guards against confused handlers.

Patch-set: 3
Reviewer: Gerrit User 1002815 <1002815@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1