npcx: i2c: change i2c settings for better timing.
The setup time of original i2c setting is closed to the margin of the
spec. This CL improves the setup time with a new setting. (The setup
time now is about 482 ns at 15 MHz.)
This CL also removes the timing settings of higher i2c source clock
frequencies which are not used so far to save the code size.

BRANCH=none
BUG=b:38217035
TEST=No build error for make buildall; run stress test with i2cxfer and
i2cscan on gru.

Change-Id: I5428a7dab1d935fd428ee9012604813e752cead8
Signed-off-by: CHLin <CHLIN56@nuvoton.com>