Update patch set 4
Patch Set 4:

The settings are covering worst case scenario. I.e., max rise and fall times as defined by the I2C standard (and are, of course, different according to the bus speed). 
We reviewed the I2C settings following the marginal setup that was measured on Bob (marginal but within spec) and we found that previous settings were based on incorrect calculations.
The settings are not intended to be board specific or channel specific so they are calculated for the worst-case scenario.

Patch-set: 4
Reviewer: Gerrit User 1165254 <1165254@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>