ps8751: assert reset for at least 1ms.
the 1ms delay isn't in the datasheets (yet), but that's what our
parade support contact recommended.

i'm applying this fix to all reef family board files, but i've only
verified this on a snappy.  the board files are pretty much clones of
each other as is the design, so that's hopefully reasonable coverage.

TEST=rebuilt, reload EC image on snappy... no ill effects noted.

BUG=b:62642003
BRANCH=firmware-reef-9042.B

Change-Id: Ic89130dad61b06e4ee4c14550e457ca92139b9da
Signed-off-by: Caveh Jalali <caveh@google.com>