tcpci: add TCPC port "initialized" state.
as it turns out, we're pretty aggressive in iterating over all TCPCs
checking for alerts when any of them generate an interrupt or their
interrupt line is low.  this can cause unfortunate behavior if the
driver hasn't initialized itself (and the chip) yet for interrupts to
be handled or we've disconnected the driver to do a TCPC firmware
update.  so, just skip ports that aren't in the "initialized" state.

note: there seems to be a quirk with the ps8751 in that it holds its
ALERT# (interrupt) line low during firmware update.  this line is
supposed to be falling edge triggered, so it's technically not
interrupting, but since we also poll the line level, we think there's
an interrupt that isn't acutally there.

the particular test case was to unplug the PD power brick during TCPC
firmware update.  the interrupt handler would be called, accessing the
TCPC over i2c and causing all sorts of havoc.

TEST=tested with follow-up CLs and verified ps8751 firmware update works.

BUG=b:35586896
BRANCH=firmware-reef-9042.B

Change-Id: Ia3444dfd57693a0daf43ff4bb43afdb0fdac1b6f