Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Where else does board ID mismatch need to be checked?
> 
> Console commands?
> 
> Case-closed debugging configuration?
> 
> GH?
> 
> How do we ensure that all new functionality on a boardID-locked image properly checks for mismatch?
> 
> I understand the goal is to ensure the device can always accept an update even if board ID gets mismatched, but why are we more worried about this than for INFO1_MASK, where a mis-write could still lock out both images?

I feel that we are going to have much more board ID variance than the info1 mask bitfield, so this is mostly to address development setup mismatches.

As for protecting new functionality - if tpm and vendor commands do not go through, the only opening is console commands, but those are locked by default and the unlock is not possible.

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>