power: rk3399: Assert SYS_RST in S5.
This commit changes SYS_RST_L to remain asserted on the transition to
S5.  SYS_RST_L is then deasserted on the transition to S0; there is no
longer any pulse for SYS_RST_L between S5 and S0.

The existing warm reset behaviour is unchanged.

BUG=b:35647982
BRANCH=gru
TEST=Flash bob; Verify that SYS_RST_L is asserted in S0.  Verify that
it's deasserted in S0.

Change-Id: Ic9c1c9c276178c7d417044a5961c3ecb2bec5635
Signed-off-by: Aseda Aboagye <aaboagye@google.com>