eCTS: Strip board serial string
If we don't strip a serial string read from a file, TH serial doesn't
match with the one read from lsusb. It causes TH to be recognized as
a DUT.

This patch also makes error messages more helpful when DUT is not found
or multiple DUTs are connected.

BUG=none
BRANCH=none
TEST=cts/cts.py -m meta, task

Change-Id: Ifcf96089b9620ef8809a7d2924dacf67d3af419c
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>