tcpci: add TCPC port "initialized" state.
as it turns out, we're pretty aggressive in iterating over all TCPCs
checking for alerts when any of them generate an interrupt or their
interrupt line is low.  this can cause unfortunate behavior if the
driver hasn't initialized itself (and the chip) yet for interrupts to
be handled or we've disconnected the driver to do a TCPC firmware
update.  so, just skip ports that aren't in the "initialized" state.

note: there seems to be a quirk with the ps8751 in that it holds its
ALERT# (interrupt) line low during firmware update.  this line is
supposed to be falling edge triggered, so it's technically not
interrupting, but since we also poll the line level, we think there's
an interrupt that isn't acutally there.  we get away with this because
pd_exchange_status() as a 5ms delay in its polling loop to avoid
spinning.

the particular test case was to unplug the PD power brick during TCPC
firmware update.  the interrupt handler would be called, accessing the
TCPC over i2c and causing all sorts of havoc.

TEST=tested with follow-up CLs and verified ps8751 firmware update
	works on electro.

BRANCH=none

BUG=b:35586896

Change-Id: I880cff49e0e9637256efa9003bcc46274536e631
Signed-off-by: Caveh Jalali <caveh@google.com>