watchdog: Don't discard irqprio data due to CONFIG_LTO
Don't discard irqprio data when the IRQ_PRIORITY macro is used directly
(for watchdog / watchdog timer).

This change is probably a NOP for all platforms, since the power-on
default for the IRQ prio register seems to be zero, which is the same
priority we're setting in our direct use of IRQ_PRIORITY.

BUG=chromium:634701
BRANCH=None
TEST=Verify 'prio_44' entry exists in irqprio section by checking
ec.RO.map on kevin.

Change-Id: I6c2f89806b170acb294523d6de1d468fe73925cf
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: ceb3e318c8c6fe15129ade47b169c3c03c729fe1
Original-Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Original-Change-Id: Idaffc484a2ce4749c18212f179b3951ff570aed0
Original-Reviewed-on: https://chromium-review.googlesource.com/545201
Original-Commit-Ready: Shawn N <shawnn@chromium.org>
Original-Tested-by: Shawn N <shawnn@chromium.org>
Original-Reviewed-by: Vincent Palatin <vpalatin@chromium.org>