eve: Check for left/right LED auto control
The Eve board defines left and right LEDs, but the LED control loop is
checking for auto control to be disabled on the power/battery LED IDs.

This is breaking factory testing for LED colors when the battery level
is low because it is flashing RED and ignoring the colors that the
factory software attempts to set.

BUG=b:63049012
BRANCH=eve
TEST=manual testing on Eve with low battery

Change-Id: Icf592cfdbe081bb507913b6f7f4df5f70f00c99e
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 3fdf85ece7c14d5fd813783d5fa67e2558ea9ecd
Original-Change-Id: I65b822111b3054a5994d71ec4e2a089c487acf97
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/549740
Original-Reviewed-by: Todd Broch <tbroch@chromium.org>