Update patch set 4
Patch Set 4:

Establishing correctness for these algorithms is very difficult. Doubly so for assembly implementations. It generally looks like a line-by-line annotation of the source code establishing bounds for each variable as they're updated to ensure a lack of carry bugs.

We don't have the review resources for that I'm afraid. We are funding some formal-methods work in this space, but it's not applicable to this case yet.

However, since this implementation is by Peter Schwabe, we can have a fair degree of confidence by reputation. (Also, this is probably based on the original qhasm sources that generated the C code.) Doing the 1M test from the RFC is also a good idea to catch integration problems.

Patch-set: 4
Reviewer: Gerrit User 1002270 <1002270@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>