Add support for reporting device events
In order to report specific wake events from differernt devices
add a host command that allows setting device event mask, and
triggering a host event when that device event is set.

This is done as a separate command and mask because we are running
out of host events, and it takes over the unused thermal overload
event that was never used in EC or BIOS.

The first use case for this is platforms that have AP wake events
that go to the EC, for instance devices that use Deep S3 and have
a limited set of wake pins. (such as Eve)

This allows the AP to determine the exact wake source for an event
so it can be logged and acted on by the AP if necessary.

BUG=b:36024430
BRANCH=eve
TEST=manual testing on eve with trackpad and dsp wake events

Change-Id: I644226eb9056fadf110e6454174c28ef977f6f02
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: bbb759ceaa843f548f90c35d1668e17c8879bad3
Original-Change-Id: I48d94014c00dc1dad098ab96af0ddc7860229762
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/555632
Original-Reviewed-by: Scott Collyer <scollyer@chromium.org>