vboot: Verify and jump to RW_A or RW_B
This patch gives EC the capability of verifying and jumping to
RW_A or RW_B. EC tries the slot stored in a persistent storage
(e.g. BBRAM). If verification fails due to invalid contents, EC
tries the other slot. AP's expectation and its reaction to the
state of the slots and the currently running image are summarized
below.

         |             AP's Expectation
         |         X                          X'
---------+-----------------------------------------------------
SLOT_A=X |      proceed              write X' to B, reboot to B
SLOT_B=X |      proceed              write X' to A, reboot to A
---------+-----------------------------------------------------
SLOT_A=X'|     reboot to B                  proceed
SLOT_B=X |      proceed                   reboot to A
---------+-----------------------------------------------------
SLOT_A=X |       proceed                  reboot to B
SLOT_B=X'|     reboot to A                  proceed
---------+-----------------------------------------------------
SLOT_A=X'|write X to B, reboot to B  write X' to B, reboot to B
SLOT_B=X'|write X to A, reboot to A  write X' to A, reboot to A

BUG=b:38462249
BRANCH=none
TEST=make buildall

Change-Id: I51e3abd4d9af44ab3d531561cb9bfa2e8d775f6a
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>