Fizz: Remove forced system lock
This patch removes the switch which force the system to be locked.
With this change, if sw write protect is enabled (thus system is
locked), PD negotiation will be prevented in the RO image. As a
consequence, the system will stay in G3 until 20W or more power is
supplied.

In case PD negotiation is required, EC will try to verify one of
the two RW images and jump to it if it's valid. Once EC gets enough
power after PD negotiatiation, it will power on the AP.

BUG=b:38462249
BRANCH=none
TEST=Lock Fizz and boot it on type-c adapter. Verify RW image is
successfully verified and executed.

Change-Id: Id8255c5c8e6af93bda3fd4de079008561f46e14c
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>