Fizz: Lock system
This patch removes CONFIG_SYSTEM_UNLOCKED, which forces the system
to be unlocked. With this change, if SW write-protect is enabled,
PD negotiation will be blocked in the RO image. As a consequence,
the system will stay in G3 until 20W or more power is supplied.

In case signed RW images are bundled, EC will try to verify one of
the images and jump to it. EC-RW will continue to negotiate more
power from a type-c charger. Once EC gets enough power, it will
power on the AP.

BUG=b:38462249
BRANCH=none
TEST=Lock Fizz and boot it on type-c adapter. Verify RW image is
successfully verified and executed.

Change-Id: Id8255c5c8e6af93bda3fd4de079008561f46e14c
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>