Update patch set 1
Patch Set 1:

(2 comments)

If I understand correctly, we see this problem on lili only and not other boards with the same isl9237 part. Is it a sign that input current regulation is broken / inaccurate on lili? Why is it the case?

Patch-set: 1
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>