usb_updater: display board ID information from RW headers
This patch adds code to interpret the contents of the Cr50 image's RW
headers board ID fields.

Board ID fields are added to the RW and RW_B version strings, as a
three field colon separated value.

The actual board ID is likely to be a 4 symbols ASCII string, if it
is, the 4 ASCII characters are displayed. If it is not - the 8
character hex value is printed. The rest of the fields are displayed
as 8 characters hex values.

BRANCH=none
BUG=b:62294740
TEST=tried the new option on various Cr50 images:
  $ ./extra/usb_updater/usb_updater -b build/cr50/ec.bin
  read 524288(0x80000) bytes from build/cr50/ec.bin
  RO_A:4919.0.0 RW_A:0.0.21[00000000:00000000:00000000] ...
  $ ./extra/usb_updater/usb_updater  -b cr50.bin.dev
  read 524288(0x80000) bytes from cr50.bin.dev
  RO_A:0.0.10 RW_A:0.0.21[XXXR:00000004:00000001]...

Change-Id: I5a92a600d24e4a7d6d615f256b5979414e883d77
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>