Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > a typical use of the .failedboard directory:
> > 
> > # make buildall -j
> > // if this fails:
> > # for b in $(ls .failedboards); do rm -rf build/$b; done
> > # repo sync
> > # make  buildall -j
> > 
> > // if this fails again - see what's wrong.
> > 
> > this patch breaks this workflow, it would be great to preserve it.
> 
> https://chromium-review.googlesource.com/563643
> 
> Theoretically, repo sync && make buildall -j should work without removing the directory. The root cause is that our Makefile isn't completely aware of all the dependencies. So, when repo sync updates some files, we get different results with incremental builds (thus people tend to do rm -rf). 
> 
> Anyway, CL 563643 should help you preserve your workflow with a slight modification.

repo sync alone often does not help because when an .h file is deleted, it is still hanging there in .d files in build/, and this is where make chokes.

But is any case, what prevents you from creating in .failedboards a file matching the directory name in ./build ? Or is it that more than one board get mapped to the same ./build subdirectory?

Patch-set: 6
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>