cr50: Always read LONG_LIFE_SCRATCH1 for brdprop.
The `brdprop` command simply returned board_properties, however,
board_properties is just a cache of the properties from init time.  When
the `brdprop` command was issued, it would only return the cached value.

This commit changes the functionality of the `brdprop` command to always
query the LONG_LIFE_SCRATCH1 register.

BUG=b:63456150
BRANCH=cr50
TEST=Flash bob.  Unplug and plug in battery, verify that
BOARD_WP_ASSERTED is set when running the `brdprop` command.

Change-Id: I0a7feb5e63f3e528b554ed0820924ac0705f5517
Signed-off-by: Aseda Aboagye <aaboagye@google.com>